requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
Deprecated==1.2.14
dj-database-url==2.1.0
Django==4.2.7
django-htmx==1.17.0
frozenlist==1.4.0
idna==3.4
Markdown==3.5.1
multidict==6.0.4
openai==0.28.1
packaging==23.2
psycopg2==2.9.9
pycparser==2.21
PyGithub==2.1.1
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-frontmatter==1.0.0
python-slugify==8.0.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
whitenoise==6.6.0
wrapt==1.15.0
yarl==1.9.2
